Ideas for Extreload tab reloading for Manifest V3
I was having a hard time jiggering the Extreload code to get tab
reloading working with Manifest V3 extensions. This is an experiment
using a better API to interact with the Chrome DevTools Protocol. It
does correctly reload a Manifest V3 extension's current tab, by sleeping
briefly to wait for the service worker to reinstall, then executing the
tab reload call.