swextreload: Debugging service worker inactive
I was getting an error sending to a null `target`. At first I thought I
should send the tab reload message to the last extension instead of the
first extension. But I later discovered that the inactive service worker
isn't even in the list of targets from `Target.getTargets`. Not sure
what to about this one. Is it possible to wake up the service worker
since we know its ID/URL?