swextreload: Ideas for problems sending messages
I'm sometimes having trouble sending messages to the extension pages.
Trying out different things here but nothing appears to be cutting it
just yet. It seems like there are times when we just don't get a
response back or can't connect to a target for whatever reason, either
in the initial extension reload step or in the tab reload.

Also, since the delay for tab reloading appears to only be relevant for
Manifest V3 extensions, I added a condition before sleeping.